% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{duality_gaps}
\alias{duality_gaps}
\title{Get Duality Gaps}
\usage{
duality_gaps(fit, family, standardize, x, y)
}
\arguments{
\item{fit}{the resulting fit}

\item{family}{the loss function}

\item{x}{design matrix}

\item{y}{response vector}
}
\value{
primals, duals, and relative duality gaps
}
\description{
Get Duality Gaps
}
