% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_gaps}
\alias{check_gaps}
\title{Get Duality Gaps}
\usage{
check_gaps(fit, family, standardize, x, y, tol_gap = 1e-04)
}
\arguments{
\item{fit}{the resulting fit}

\item{family}{the loss function}

\item{x}{design matrix}

\item{y}{response vector}
}
\value{
primals, duals, and relative duality gaps
}
\description{
Get Duality Gaps
}
